+++ MyLibraryCPP/clsSquare.h
@@ -7,40 +7,40 @@ using namespace std;
 class clsSquare
 {
 private:
-	float _SquareSide;
+	double _SquareSide;
 
 public:
 	static constexpr double PI = 3.14159;
 
-	clsSquare(float  squareSide)
+	clsSquare(double  squareSide)
 	{
 		_SquareSide = squareSide;
 	}
 
-	float  getSquareSide()
+	double  getSquareSide()
 	{
 		return _SquareSide;
 	}
 
-	void setSquareSide(float  squareSide)
+	void setSquareSide(double  squareSide)
 	{
 		_SquareSide = squareSide;
 	}
 
-	static float CalculateCircleAreaInscribedInSquare(float SquareSide)
+	static double CalculateCircleAreaInscribedInSquare(double SquareSide)
 	{
 		return PI * pow(SquareSide, 2) / pow(2, 2);
 	}
 
-	static float CalculateCircleAreaInscribedInSquare(clsSquare Square)
+	static double CalculateCircleAreaInscribedInSquare(clsSquare Square)
 	{
 		return CalculateCircleAreaInscribedInSquare(Square._SquareSide);
 	}
 
-	float CalculateCircleAreaInscribedInSquare()
+	double CalculateCircleAreaInscribedInSquare()
 	{
 		return CalculateCircleAreaInscribedInSquare(_SquareSide);
 	}
 
-	__declspec(property(get = getRadius, put = setRadius))	float  squareSide;
+	__declspec(property(get = getRadius, put = setRadius))	double  squareSide;
 };
